refactor(morpho-sdk): unify Blue funding prefix in buildAssetFundingActions
Rename buildRepayFundingActions -> buildAssetFundingActions and generalize it
to the shared native-wrap + ERC-20-pull prefix that every bundled Blue funding
path emits before its Morpho call. The helper now takes `asset: Address`
instead of `marketParams`, and drops the redundant explicit erc20TransferFrom
branch: getTokenRequirementActions already emits that action itself when no
permit/permit2 signature is supplied.

Wired into all 5 Blue builders (previously 2), making it the single source of
truth for the funding prefix:
- blueSupply / blueSupplyCollateral / blueSupplyCollateralBorrow
- blueRepay / blueRepayWithdrawCollateral

Behavior-preserving: each builder emits a byte-identical Action[] for all
inputs, so the encoded bundles are unchanged (76 unit + 52 fork e2e green).
Vault deposits are intentionally left out — they throw the distinct public
NativeAmountOnNonWNativeVaultError, so folding them in would change that error
contract.

The two repay unit tests that asserted getTokenRequirementActions was not
called without a signature now assert the unified routing (called once, with
requirementSignature undefined); the observable erc20TransferFrom is proven by
the fork round-trips.

Internal @internal helper, public surface unchanged: no changeset.
